+++ CHANGELOG.md
@@ -5,6 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.41.0] - Unpublished
+
+### Added
+
+- [enhancement](https://github.com/FrankC01/pysui/issues/176) Support for address/kp aliases
+
+### Fixed
+
+### Changed
+
+- [enhancement](https://github.com/FrankC01/pysui/issues/176) Added alias optional string to:
+  * SuiConfig.create_new_keypair_and_address
+  * SuiConfig.recover_keypair_and_address
+  * SuiConfig.add_keypair_from_keystring
+- Added optional 'make_active' flag to:
+  * SuiConfig.create_new_keypair_and_address
+  * SuiConfig.recover_keypair_and_address
+
+
+### Removed
+
 ## [0.40.1] - Unpublished
 
 ### Added